refactor(store): move store settings fetches to server-only data module
